Spark streaming fixed

+++ cryptotrade-spark/src/main/scala/dmitrypukhov/cryptotrade/model/BidAsk.scala
@@ -2,4 +2,4 @@ package dmitrypukhov.cryptotrade.model
 
 import java.time.LocalDateTime
 
-case class BidAsk(time: LocalDateTime, symbol: String, bid: BigDecimal,bidQty: Int, ask: BigDecimal, askQty:Int)
+case class BidAsk(time: LocalDateTime, symbol: String, bid: BigDecimal,bidQty: BigDecimal, ask: BigDecimal, askQty:BigDecimal)
